Moved `ReviewData` to a separate class file
